add binary sensor platform
